% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/and_vertebrates.R
\docType{data}
\name{and_vertebrates}
\alias{and_vertebrates}
\title{and_vertebrates}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 32209 rows and 18 columns.
}
\source{
{AND LTER}
\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-and.4027.14}
}
\usage{
and_vertebrates
}
\description{
Aquatic Vertebrate Population Study in Mack Creek, Andrews Experimental Forest, 1987 to present.
}
\keyword{datasets}
