% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anole_mark.R
\docType{data}
\name{anole_mark}
\alias{anole_mark}
\title{anole_mark}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 543 rows and 16 columns.
}
\source{
{LEF LTER}
\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-luq.5.347700}
}
\usage{
anole_mark
}
\description{
Jornada Basin LTER anole mark and resight study.
}
\keyword{datasets}
