% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnz_peatland.R
\docType{data}
\name{bnz_peatland}
\alias{bnz_peatland}
\title{bnz_peatland}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 41 rows and 39 columns.
}
\source{
{BNZ LTER}
\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-bnz.468.14}
}
\usage{
bnz_peatland
}
\description{
Alaskan Peatland Experiment: Community structure and productivity data for 2007-2010 II - Species Abundance
}
\keyword{datasets}
