% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnz_seedfall.R
\docType{data}
\name{bnz_seedfall}
\alias{bnz_seedfall}
\title{bnz_seedfall}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4646 rows and 15 columns.
}
\source{
{BNZ LTER}
\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-bnz.14.28}
}
\usage{
bnz_seedfall
}
\description{
Yearly Seedfall Summary from 1985 to Present in the Bonanza Creek Experimental Forest near Fairbanks, Alaska.
}
\keyword{datasets}
