% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf_soil.R
\docType{data}
\name{hf_soil}
\alias{hf_soil}
\title{hf_soil}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 27 rows and 8 columns.
}
\source{
{HF LTER}
\url{https://portal.edirepository.org/nis/mapbrowse?packageid=knb-lter-hfr.343.1}
}
\usage{
hf_soil
}
\description{
Soil Carbon and Nitrogen at the Harvard Farm at Harvard Forest since 2015
}
\keyword{datasets}
