% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nes_diet.R
\docType{data}
\name{nes_diet}
\alias{nes_diet}
\title{nes_diet}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1409 rows and 22 columns.
}
\source{
{NES LTER}
\url{https://portal.edirepository.org/nis/metadataviewer?packageid=knb-lter-nes.2.2}
}
\usage{
nes_diet
}
\description{
Diet Composition for Small Pelagic Fishes across the Northeast U.S. Continental Shelf from 2013-2015
}
\keyword{datasets}
